Added to the single generator case
The text there was only for idempotent generators, which won't work in
general.  Happily it's OK for a Gaussian distribution with any
generator.  In fact, it seems that you get the analytically continued
version of the charateristic function for a Gaussian.  I cannot help but
wonder if that works for all distributions!

(Help me if you can prove that!)