Drastic change in penultimate section
I have kept the original text but commented it out.  The issues I had
with much of it are made in previous updates.  I really couldn't come to
solve any of those issues.  So I did a slight rewrite.  Given the length
of this and the main idea being stated fairly precisely, I think vague
discussions like what I've added are OK.  We can always resurrect any of
the previous text if need be.