Implamented some of the suggested changes
I have added in a bit in section 2 about the types of errors (but haven't meantioned mode mismatch)  and about errors in the encoding beam splitters.
I have not included the "simple example to help clarify the argument" in section 2 as I am not really sure where it would go.

I have also continued to remove the comments left in and have tried to unify the manner in which we refer to equations.